refactor: relocate provider helper to host support
