+++ examples/integrations/litellm/basic.py
@@ -18,7 +18,7 @@ from typing import TypedDict, cast
 
 from context_compiler import State, get_policy_items, get_premise_value
 from context_compiler.engine import Engine
-from examples._provider import print_startup_config, resolve_provider_config
+from host_support.provider_mode import print_startup_config, resolve_provider_config
 
 logger = logging.getLogger(__name__)
 # Example-only in-memory checkpoint store.
