+++ host_support/__init__.py
@@ -1,0 +1,1 @@
+"""Host-side shared helpers."""
